backend/pyproject.toml
[project]
name = "taxgpt-backend"
version = "0.1.0"
description = "FastAPI service powering the TaxGPT hackathon agent."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "TaxGPT Team", email = "engineering@taxgpt.dev" }]
dependencies = [
    "fastapi>=0.115.5,<1.0.0",
    "uvicorn[standard]>=0.30.6,<1.0.0",
    "sqlmodel>=0.0.21",
    "psycopg[binary]>=3.2.1",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    "structlog>=24.4.0",
    "httpx>=0.27.2",
    "rich>=13.9.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.8",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.11.2",
    "types-requests>=2.32.0.20241016",
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--strict-markers --disable-warnings --maxfail=1"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "N",
]
fixable = ["ALL"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
strict_optional = true

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.8",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.11.2",
    "types-requests>=2.32.0.20241016",
]

